Added login_required annotations to all users files to ensure only authorised, logged-in users can access such pages and therefore no 'anonoymous' users can use  and access the lottery web app's restricted pages.
Also fixed bug where field 'username' was mis-spelt and was meant to be the 'email' field. 'pinkey' was also misspelled and meant to be pin_key. This error was present in the login.html and views.py

Added GET and POST methods to lottery method in lottery views.py.
Also added logout function to views.py.

Also corrected pin length - was meant to be 32 characters long (on registration)