Fixed error in redirect link in users\views.py where incorrect page link was specified.
Ensured all played draws displayed on lottery page - lottery\views.py - are only displayed for the currently logged in user.

Passed the firstname variable of the current user to the index page in app.py.

Amended the text displayed in index.html to ensure it welcomes the specific user who is logged in by their name, and if there is no user logged in (an anonymous user is using the web app) it will direct them to login or register.